+++ src/main/java/me/ilyamirin/anthophila/hands/Hand.java
@@ -10,7 +10,7 @@ public interface Hand {
 
     long size() throws IOException;
 
-    void write(int position, ByteBuffer bufferToWrite) throws IOException;
+    void write(Long position, ByteBuffer bufferToWrite) throws IOException;
 
-    ByteBuffer read(int position, int size) throws IOException;
+    ByteBuffer read(Long position, int size) throws IOException;
 }
